# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: AMX — Account Management Server API
  version: 0.1.0
  summary: Tenant, account, server and assignment management for AMX.
  description: |
    Management plane of AMX (Account Management & eXchange). Everything is scoped
    to a tenant: accounts, servers and assignments all belong to exactly one, and
    the tenant boundary is a structural database invariant, not an application
    check (§5.1 composite foreign keys). A cross-tenant assignment is rejected by
    the database itself.

    **Credential secrets are write-only.** `accounts.encrypted_secret` holds a
    complete OAuth credential set (access + refresh + expiry + scopes + account
    metadata, §5.5) encrypted with `AMX_ENCRYPTION_KEY`. No response on this API
    ever returns it — account responses expose only masked metadata
    (`secretMasked`, `credentialExpiresAt`). Tokens must never be logged (§7).

    **Action endpoints** use the `:verb` suffix (`…/{id}:deliver`). They are
    non-RESTful on purpose: they move an assignment through the state machine of
    §5.2, or issue a non-state command (`:switch-now`, `:switch-mode`,
    `:refresh-usage`) that changes no assignment state. All of them are
    asynchronous — the response reports the state AMS has recorded, and
    convergence arrives later as a `CommandAck` over the gRPC session.
  license:
    name: Proprietary
  contact:
    name: AMX
    url: https://amx.io
servers:
  - url: https://{host}/api/v1
...
    variables:
...
security:
  - bearerAuth: []
...
paths:
...
  /tenants/{tenantId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [tenants]
      operationId: getTenant
      summary: Get a tenant
      responses:
        '200':
          description: Tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [tenants]
      operationId: updateTenant
      summary: Update a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantUpdate'
      responses:
        '200':
          description: Updated tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [tenants]
      operationId: deleteTenant
      summary: Delete a tenant
      description: Rejected while the tenant still owns accounts, servers, admins or non-detached assignments.
      responses:
        '204':
          description: Deleted.
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ---------------------------------------------------------------------------
  # Accounts
  # ---------------------------------------------------------------------------
...
  /tenants/{tenantId}/usage/langfuse:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [usage]
      operationId: getLangfuseUsage
      summary: Langfuse token roll-up for a tenant
      description: |
        Per-day token totals compacted from the Langfuse Metrics API by the
        server-side metrics sweep, split into a model axis and a user axis. Both
        `from`/`to` bounds are inclusive UTC days and default to the trailing week
        ending today. A tenant other than the configured `AMX_LANGFUSE_TENANT_ID`,
        or a range with no data, is an empty 200. `cacheReadTokens` /
        `cacheCreationTokens` are always 0 today — the Metrics API exposes no
        cache-token measure.
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Model-axis and user-axis token roll-up rows for the range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LangfuseUsageResponse'
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Administrator session token. Tenant RBAC is enforced per request (§7).
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: No such resource in this tenant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflicts with current state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      description: RFC 9457 problem detail.
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          description: Stable machine-readable code, e.g. `assignment.invalid_transition`.
          type: string
...
    TenantStatus:
      type: string
      enum: [active, suspended]
    Tenant:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/TenantStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          $ref: '#/components/schemas/TenantStatus'
...
    LangfuseModelRow:
      type: object
      required: [day, model, inputTokens, outputTokens, cacheReadTokens, cacheCreationTokens, totalTokens, observations]
      properties:
        day:
          type: string
          format: date
        model:
          type: string
          description: providedModelName, or "unknown" when Langfuse reports it null.
        inputTokens:
          type: integer
          format: int64
        outputTokens:
          type: integer
          format: int64
        cacheReadTokens:
          type: integer
          format: int64
          description: Always 0 today — the Metrics API exposes no cache-token measure.
        cacheCreationTokens:
          type: integer
          format: int64
          description: Always 0 today — the Metrics API exposes no cache-token measure.
        totalTokens:
          type: integer
          format: int64
        observations:
          type: integer
    LangfuseUserRow:
      type: object
      required: [day, userId, totalTokens, observations]
      properties:
        day:
          type: string
          format: date
        userId:
          type: string
          description: The account email fixed as the Metrics API userId filter.
        totalTokens:
          type: integer
          format: int64
        observations:
          type: integer
    LangfuseUsageResponse:
      type: object
      required: [modelRows, userRows]
      properties:
        modelRows:
          type: array
          items:
            $ref: '#/components/schemas/LangfuseModelRow'
        userRows:
          type: array
          items:
            $ref: '#/components/schemas/LangfuseUserRow'
        uiUrl:
          type: string
          nullable: true
          description: Console deep-link base (AMX_LANGFUSE_UI_URL, else the API base, else null).